requirements.txt
streamlit
pandas
requests
matplotlib
seaborn
plotly.express
requests
matplotlib.pyplot